+++ packages/sdk/test/end-to-end/StreamRegistry.test.ts
@@ -1,12 +1,11 @@
 import 'reflect-metadata'
 
 import { fetchPrivateKeyWithGas, randomEthereumAddress, randomUserId } from '@streamr/test-utils'
-import { EthereumAddress, collect, toEthereumAddress, toStreamID, waitForCondition } from '@streamr/utils'
+import { EthereumAddress, collect, toEthereumAddress, toStreamID, until } from '@streamr/utils'
 import { Wallet } from 'ethers'
 import { CONFIG_TEST } from '../../src/ConfigTest'
 import { Stream } from '../../src/Stream'
 import { StreamrClient } from '../../src/StreamrClient'
-import { until } from '../../src/utils/promises'
 import { createRelativeTestStreamId, createTestStream } from '../test-utils/utils'
 
 const TIMEOUT = 20000
@@ -94,7 +93,7 @@ describe('StreamRegistry', () => {
                 const streamIds = onStreamCreated.mock.calls.map((c) => c[0].streamId)
                 return streamIds.includes(stream.id)
             }
-            await waitForCondition(() => hasBeenCalledFor(validStream))
+            await until(() => hasBeenCalledFor(validStream))
             client.off('streamCreated', onStreamCreated)
             expect(onStreamCreated).toHaveBeenCalledWith({
                 streamId: validStream.id,
@@ -236,10 +235,10 @@ describe('StreamRegistry', () => {
         }, TIMEOUT)
     })
 
-    describe('update', () => {
+    describe('setMetadata', () => {
         it('happy path', async () => {
             const description = `description-${Date.now()}`
-            await createdStream.update({
+            await createdStream.setMetadata({
                 description
             })
             await until(async () => {
@@ -261,7 +260,7 @@ describe('StreamRegistry', () => {
         it('happy path', async () => {
             const props = { id: createRelativeTestStreamId(module) }
             const stream = await client.createStream(props)
-            await stream.delete()
+            await client.deleteStream(stream.id)
             await until(async () => {
                 try {
                     await client.getStream(stream.id)