+++ packages/utils/test/until.test.ts
@@ -1,9 +1,9 @@
-import { waitForCondition } from '../src/waitForCondition'
+import { until } from '../src/until'
 
-describe('waitForCondition', () => {
+describe('until', () => {
     describe('given conditionFn that returns boolean primitives', () => {
         it('resolves immediately if conditionFn returns true from the get-go', (done) => {
-            waitForCondition(() => true)
+            until(() => true)
                 .then(done)
                 .catch(() => done(new Error('timed out')))
         })
@@ -11,15 +11,15 @@ describe('waitForCondition', () => {
         it('resolves eventually when conditionFn returns true', (done) => {
             let cbReturnValue = false
             setTimeout(() => cbReturnValue = true, 50)
-            waitForCondition(() => cbReturnValue, 5000, 10)
+            until(() => cbReturnValue, 5000, 10)
                 .then(done)
                 .catch(() => done(new Error('timed out')))
         })
 
         it('rejects if conditionFn does not return true within timeout', (done) => {
             const pollCb = () => false
-            waitForCondition(pollCb, 50, 5).catch((err) => {
-                expect(err.message).toEqual('waitForCondition: timed out before "() => false" became true')
+            until(pollCb, 50, 5).catch((err) => {
+                expect(err.message).toEqual('until: timed out before "() => false" became true')
                 done()
             })
         })
@@ -29,8 +29,8 @@ describe('waitForCondition', () => {
             setTimeout(() => {
                 abortController.abort()
             }, 100)
-            waitForCondition(() => false, 5000, 10, abortController.signal).catch((err) => {
-                expect(err.message).toEqual('waitForCondition: aborted before "() => false" became true')
+            until(() => false, 5000, 10, abortController.signal).catch((err) => {
+                expect(err.message).toEqual('until: aborted before "() => false" became true')
                 done()
             })
         })
@@ -39,7 +39,7 @@ describe('waitForCondition', () => {
     describe('given conditionFn that returns promisified booleans (i.e. Promise<boolean>)', () => {
         it('resolves immediately if conditionFn returns (promisified) true from the get-go', async () => {
             const fn = jest.fn().mockResolvedValue(true)
-            await waitForCondition(fn)
+            await until(fn)
             expect(fn).toBeCalledTimes(1)
         })
 
@@ -47,19 +47,19 @@ describe('waitForCondition', () => {
             const fn = jest.fn()
                 .mockResolvedValueOnce(false)
                 .mockResolvedValueOnce(true)
-            await waitForCondition(fn)
+            await until(fn)
             expect(fn).toBeCalledTimes(2)
         })
 
         it('rejects if conditionFn keeps returning (promisified) false within timeout', async () => {
             const fn = () => Promise.resolve(false)
-            await expect(waitForCondition(fn, 50, 10)).rejects
-                .toThrow('waitForCondition: timed out before "() => Promise.resolve(false)" became true')
+            await expect(until(fn, 50, 10)).rejects
+                .toThrow('until: timed out before "() => Promise.resolve(false)" became true')
         })
 
         it('rejects immediately if conditionFn returns rejected promise from the get-go', async () => {
             const error = new Error('mock')
-            await expect(waitForCondition(() => Promise.reject(error))).rejects.toThrow(error)
+            await expect(until(() => Promise.reject(error))).rejects.toThrow(error)
         })
 
         it('rejects eventually if conditionFn returns rejected promise and no (promisifed) true was encountered', async () => {
@@ -67,15 +67,15 @@ describe('waitForCondition', () => {
             const fn = jest.fn()
                 .mockResolvedValueOnce(false)
                 .mockRejectedValueOnce(error)
-            await expect(waitForCondition(fn))
+            await expect(until(fn))
                 .rejects
                 .toThrow(error)
         })
 
         it('rejects if conditionFn returns promise that does not settle within timeout', async () => {
-            await expect(waitForCondition(() => new Promise(() => {}), 100, 10))
+            await expect(until(() => new Promise(() => {}), 100, 10))
                 .rejects
-                .toThrow('waitForCondition: timed out before "() => new Promise(() => { })" became true')
+                .toThrow('until: timed out before "() => new Promise(() => { })" became true')
         })
 
         it('rejects if conditionFn does not return true before abort signalled', (done) => {
@@ -83,8 +83,8 @@ describe('waitForCondition', () => {
             setTimeout(() => {
                 abortController.abort()
             }, 100)
-            waitForCondition(() => Promise.resolve(false), 5000, 10, abortController.signal).catch((err) => {
-                expect(err.message).toEqual('waitForCondition: aborted before "() => Promise.resolve(false)" became true')
+            until(() => Promise.resolve(false), 5000, 10, abortController.signal).catch((err) => {
+                expect(err.message).toEqual('until: aborted before "() => Promise.resolve(false)" became true')
                 done()
             })
         })
@@ -93,16 +93,16 @@ describe('waitForCondition', () => {
     it('rejects immediately if given pre-aborted signal', (done) => {
         const abortController = new AbortController()
         abortController.abort()
-        waitForCondition(() => true, 5000, 1, abortController.signal).catch((err) => {
-            expect(err.message).toEqual('waitForCondition: aborted before "() => true" became true')
+        until(() => true, 5000, 1, abortController.signal).catch((err) => {
+            expect(err.message).toEqual('until: aborted before "() => true" became true')
             done()
         })
     })
 
     it('can provide contextual information on rejection', (done) => {
         const pollCb = () => false
-        waitForCondition(pollCb, 50, 5, undefined, () => 'a was 5, expected 10').catch((err) => {
-            expect(err.message).toEqual('waitForCondition: timed out before "() => false" became true' +
+        until(pollCb, 50, 5, undefined, () => 'a was 5, expected 10').catch((err) => {
+            expect(err.message).toEqual('until: timed out before "() => false" became true' +
                 '\na was 5, expected 10')
             done()
         })